+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod computer;
+pub mod osero;
+pub mod test;
