+++ src/osero.rs
@@ -1,0 +1,2 @@
+pub mod game;
+pub mod util;
