+++ src/test.rs
@@ -1,0 +1,4 @@
+#[cfg(test)]
+mod game_test;
+#[cfg(test)]
+mod util_test;
